# ⚡️ Speed up function `_map_usage` by 172%